DLQConsumer working FE
