+++ demo/server/routes/strategy.ts
@@ -1,17 +1,20 @@
 import * as express from 'express';
 import kafkaController from '../controllers/kafkaController';
-import strategyController from '../controllers/strategyController';
+import failfastController from '../controllers/failfastController';
+import DLQController from '../controllers/DLQController'
 const router = express.Router();
 
-router.post('/failfast', strategyController.failfast, (req, res) => {
-  return res.status(200).json(res.locals.error);
+router.post('/failfast', 
+  failfastController.failfast,  
+  (req, res) => {
+    return res.status(200).json(res.locals.error);
 });
 
 router.post(
   '/dlq',
   kafkaController.makeClient,
-  strategyController.dlqProduce,
-  strategyController.dlqConsume,
+  DLQController.dlqProduce,
+  DLQController.dlqConsume,
   (req, res) => {
     // return res.status(200).json(res.locals.messages);
   }
